improve implicit treap: support large index

+++ internal/types.hpp
@@ -7,7 +7,7 @@ namespace lib {
 namespace internal {
 
 
-using size_t = std::int32_t;
+using size_t = std::int64_t;
 
 using int128_t = __int128_t;
 using uint128_t = __uint128_t;
